+++ src/lib/converter/plugins/SourcePlugin.ts
@@ -1,11 +1,6 @@
 import * as ts from "typescript";
 
-import {
-    Reflection,
-    ProjectReflection,
-    DeclarationReflection,
-} from "../../models/reflections/index";
-import { SourceFile } from "../../models/sources/index";
+import type { Reflection } from "../../models/reflections/index";
 import { Component, ConverterComponent } from "../components";
 import { Converter } from "../converter";
 import type { Context } from "../context";
@@ -13,7 +8,6 @@ import { BindOption } from "../../utils";
 import { isNamedNode } from "../utils/nodes";
 import { getCommonDirectory, normalizePath } from "../../utils/fs";
 import { relative } from "path";
-import * as assert from "assert";
 import { SourceReference } from "../../models";
 
 /**
@@ -25,15 +19,9 @@ export class SourcePlugin extends ConverterComponent {
     readonly disableSources!: boolean;
 
     /**
-     * A map of all generated {@link SourceFile} instances.
-     */
-    private fileMappings: { [name: string]: SourceFile } = {};
-
-    /**
      * All file names to find the base path from.
      */
     private fileNames = new Set<string>();
-    private basePath?: string;
 
     /**
      * Create a new SourceHandler instance.
@@ -44,28 +32,11 @@ export class SourcePlugin extends ConverterComponent {
             [Converter.EVENT_CREATE_DECLARATION]: this.onDeclaration,
             [Converter.EVENT_CREATE_SIGNATURE]: this.onDeclaration,
             [Converter.EVENT_RESOLVE_BEGIN]: this.onBeginResolve,
-            [Converter.EVENT_RESOLVE]: this.onResolve,
-            [Converter.EVENT_RESOLVE_END]: this.onEndResolve,
         });
     }
 
-    private getSourceFile(
-        fileName: string,
-        project: ProjectReflection
-    ): SourceFile {
-        if (!this.fileMappings[fileName]) {
-            const file = new SourceFile(fileName);
-            this.fileMappings[fileName] = file;
-            project.files.push(file);
-        }
-
-        return this.fileMappings[fileName];
-    }
-
     private onEnd() {
-        this.fileMappings = {};
         this.fileNames.clear();
-        this.basePath = void 0;
     }
 
     /**
@@ -73,20 +44,17 @@ export class SourcePlugin extends ConverterComponent {
      *
      * Attach the current source file to the {@link DeclarationReflection.sources} array.
      *
-     * @param context  The context object describing the current state the converter is in.
+     * @param _context  The context object describing the current state the converter is in.
      * @param reflection  The reflection that is currently processed.
      */
-    private onDeclaration(context: Context, reflection: Reflection) {
-        const symbol = reflection.project.getSymbolFromReflection(reflection);
-        if (!symbol || this.disableSources) {
-            return;
-        }
+    private onDeclaration(_context: Context, reflection: Reflection) {
+        if (this.disableSources) return;
 
-        for (const node of symbol.declarations || []) {
+        const symbol = reflection.project.getSymbolFromReflection(reflection);
+        for (const node of symbol?.declarations || []) {
             const sourceFile = node.getSourceFile();
             const fileName = sourceFile.fileName;
             this.fileNames.add(fileName);
-            const file = this.getSourceFile(fileName, context.project);
 
             let position: ts.LineAndCharacter;
             if (isNamedNode(node)) {
@@ -101,17 +69,12 @@ export class SourcePlugin extends ConverterComponent {
                 );
             }
 
-            if (reflection instanceof DeclarationReflection) {
-                file.reflections.push(reflection);
-            }
-
             reflection.sources ||= [];
             reflection.sources.push(
                 new SourceReference(
                     fileName,
                     position.line + 1,
-                    position.character,
-                    file
+                    position.character
                 )
             );
         }
@@ -123,39 +86,15 @@ export class SourcePlugin extends ConverterComponent {
      * @param context  The context object describing the current state the converter is in.
      */
     private onBeginResolve(context: Context) {
-        this.basePath = getCommonDirectory([...this.fileNames]);
-        for (const file of context.project.files) {
-            const fileName = (file.fileName = normalizePath(
-                relative(this.basePath, file.fileName)
-            ));
-            this.fileMappings[fileName] = file;
-        }
-    }
+        if (this.disableSources) return;
+        const basePath = getCommonDirectory([...this.fileNames]);
 
-    /**
-     * Triggered when the converter resolves a reflection.
-     *
-     * @param context  The context object describing the current state the converter is in.
-     * @param reflection  The reflection that is currently resolved.
-     */
-    private onResolve(_context: Context, reflection: Reflection) {
-        assert(this.basePath != null);
-        for (const source of reflection.sources ?? []) {
-            source.fileName = normalizePath(
-                relative(this.basePath, source.fileName)
-            );
+        for (const refl of Object.values(context.project.reflections)) {
+            for (const source of refl.sources || []) {
+                source.fileName = normalizePath(
+                    relative(basePath, source.fullFileName)
+                );
+            }
         }
-    }
-
-    /**
-     * Triggered when the converter has finished resolving a project.
-     *
-     * @param context  The context object describing the current state the converter is in.
-     */
-    private onEndResolve(context: Context) {
-        const project = context.project;
-        project.files.forEach((file) => {
-            file.reflections = file.reflections.slice();
-        });
     }
 }