+++ src/lib/models/sources/file.ts
@@ -1,8 +1,3 @@
-import * as Path from "path";
-
-import type { ReflectionGroup } from "../ReflectionGroup";
-import type { RepositoryType } from "./repository";
-import type { DeclarationReflection } from "..";
 import type { SourceReference as JSONSourceReference } from "../../serialization/schema";
 
 /**
@@ -12,14 +7,14 @@ import type { SourceReference as JSONSourceReference } from "../../serialization
  */
 export class SourceReference {
     /**
-     * A reference to the corresponding file instance.
-     */
-    file: SourceFile;
-
-    /**
      * The filename of the source file.
      */
     fileName: string;
+
+    /**
+     * The absolute filename of the source file.
+     */
+    fullFileName: string;
 
     /**
      * The number of the line that emitted the declaration.
@@ -36,16 +31,11 @@ export class SourceReference {
      */
     url?: string;
 
-    constructor(
-        fileName: string,
-        line: number,
-        character: number,
-        file: SourceFile
-    ) {
+    constructor(fileName: string, line: number, character: number) {
         this.fileName = fileName;
+        this.fullFileName = fileName;
         this.line = line;
         this.character = character;
-        this.file = file;
     }
 
     toObject(): JSONSourceReference {
@@ -55,64 +45,5 @@ export class SourceReference {
             character: this.character,
             url: this.url,
         };
-    }
-}
-
-/**
- * Exposes information about a source file.
- *
- * One may access a list of all source files through the {@link ProjectReflection.files} property or as
- * a tree structure through the {@link ProjectReflection.directory} property.
- *
- * Furthermore each reflection carries references to the related SourceFile with their
- * {@link DeclarationReflection.sources} property. It is an array of of {@link SourceReference} instances
- * containing the reference in their {@link SourceReference.file} field.
- */
-export class SourceFile {
-    /**
-     * The original full system file name.
-     */
-    fullFileName: string;
-
-    /**
-     * A trimmed version of the file name. Contains only the path relative to the
-     * determined base path.
-     */
-    fileName: string;
-
-    /**
-     * The base name of the file.
-     */
-    name: string;
-
-    /**
-     * A URL pointing to a page displaying the contents of this file.
-     */
-    url?: string;
-
-    /**
-     * The type of repository where this file is hosted.
-     */
-    repositoryType?: RepositoryType;
-
-    /**
-     * A list of all reflections that are declared in this file.
-     */
-    reflections: DeclarationReflection[] = [];
-
-    /**
-     * A grouped list of the reflections declared in this file.
-     */
-    groups?: ReflectionGroup[];
-
-    /**
-     * Create a new SourceFile instance.
-     *
-     * @param fullFileName  The full file name.
-     */
-    constructor(fullFileName: string) {
-        this.fileName = fullFileName;
-        this.fullFileName = fullFileName;
-        this.name = Path.basename(fullFileName);
     }
 }