+++ src/lib/models/sources/index.ts
@@ -1,3 +1,2 @@
-export { SourceFile } from "./file";
 export { SourceReference } from "./file";
 export { RepositoryType } from "./repository";